+++ src/engine.cpp
@@ -1,6 +1,7 @@
 #include "spsa.h"
 #include "uci.h"
 #include "tt.h"
+#include "zobrist.h"
 #include "magic.h"
 #include "bitboard.h"
 #include "nnue.h"
