+++ src/tt.cpp
@@ -1,81 +1,47 @@
-#include <random>
-#include <iostream>
-
 #include "tt.h"
 #include "move.h"
 
-uint64_t ZOBRIST_PIECE_SQUARES[2][PIECE_TYPES][64];
-uint64_t ZOBRIST_STM_BLACK;
-uint64_t ZOBRIST_NO_PAWNS;
-uint64_t ZOBRIST_CASTLING[16];
-uint64_t ZOBRIST_ENPASSENT[8];
+void TTEntry::update(uint64_t _hash, Move _bestMove, uint8_t _depth, Eval _eval, Eval _value, bool wasPv, int _flags) {
+    // Update bestMove if not MOVE_NONE
+    // Or even clear move for a different position
+    if (_bestMove != MOVE_NONE || (uint16_t)_hash != hash)
+        bestMove = _bestMove;
 
-uint64_t CUCKOO_HASHES[8192];
-Move CUCKOO_MOVES[8192];
+    if (_flags == TT_EXACTBOUND || (uint16_t)_hash != hash || depth == NO_DEPTH || _depth + 2 * wasPv + 4 > depth) {
+        hash = (uint16_t)_hash;
+        depth = _depth;
+        value = _value;
+        eval = _eval;
+        flags = (uint8_t)(_flags + (wasPv << 2)) | TT_GENERATION_COUNTER;
+    }
+}
+
+TTEntry* TranspositionTable::probe(uint64_t hash, bool* found) {
+    TTCluster* cluster = &table[index(hash)];
+    uint64_t hash16 = (uint16_t)hash;
+
+    TTEntry* replace = &cluster->entries[0];
+
+    for (int i = 0; i < CLUSTER_SIZE; i++) {
+        if (cluster->entries[i].hash == hash16 || cluster->entries[i].depth == NO_DEPTH) {
+            // Refresh generation
+            cluster->entries[i].flags = (uint8_t)(TT_GENERATION_COUNTER | (cluster->entries[i].flags & (GENERATION_DELTA - 1)));
+            *found = cluster->entries[i].hash == hash16;
+            return &cluster->entries[i];
+        }
+
+        if (i > 0) {
+            // Check if this entry would be better suited for replacement than the current replace entry
+            int replaceValue = replace->depth - ((GENERATION_CYCLE + TT_GENERATION_COUNTER - replace->flags) & GENERATION_MASK);
+            int entryValue = cluster->entries[i].depth - ((GENERATION_CYCLE + TT_GENERATION_COUNTER - cluster->entries[i].flags) & GENERATION_MASK);
+            if ((cluster->entries[i].depth == NO_DEPTH && replace->depth != NO_DEPTH) || replaceValue > entryValue)
+                replace = &cluster->entries[i];
+        }
+    }
+
+    *found = false;
+    return replace;
+}
 
 uint8_t TT_GENERATION_COUNTER = 0;
-
 TranspositionTable TT;
-
-void initZobrist() {
-    std::mt19937 rng;
-    rng.seed(934572);
-    std::uniform_int_distribution<uint64_t> dist;
-
-    for (int side = 0; side <= 1; side++) {
-        for (int i = 0; i < PIECE_TYPES; i++) {
-            for (int j = 0; j < 64; j++) {
-                ZOBRIST_PIECE_SQUARES[side][i][j] = dist(rng);
-            }
-        }
-    }
-    ZOBRIST_STM_BLACK = dist(rng);
-    ZOBRIST_NO_PAWNS = dist(rng);
-
-    ZOBRIST_CASTLING[0x0] = 0;
-    ZOBRIST_CASTLING[0x1] = dist(rng);
-    ZOBRIST_CASTLING[0x2] = dist(rng);
-    ZOBRIST_CASTLING[0x4] = dist(rng);
-    ZOBRIST_CASTLING[0x8] = dist(rng);
-    ZOBRIST_CASTLING[0x1 | 0x2] = ZOBRIST_CASTLING[0x1] ^ ZOBRIST_CASTLING[0x2];
-    ZOBRIST_CASTLING[0x1 | 0x4] = ZOBRIST_CASTLING[0x1] ^ ZOBRIST_CASTLING[0x4];
-    ZOBRIST_CASTLING[0x1 | 0x8] = ZOBRIST_CASTLING[0x1] ^ ZOBRIST_CASTLING[0x8];
-    ZOBRIST_CASTLING[0x2 | 0x4] = ZOBRIST_CASTLING[0x2] ^ ZOBRIST_CASTLING[0x4];
-    ZOBRIST_CASTLING[0x2 | 0x8] = ZOBRIST_CASTLING[0x2] ^ ZOBRIST_CASTLING[0x8];
-    ZOBRIST_CASTLING[0x4 | 0x8] = ZOBRIST_CASTLING[0x4] ^ ZOBRIST_CASTLING[0x8];
-    ZOBRIST_CASTLING[0x1 | 0x2 | 0x4] = ZOBRIST_CASTLING[0x1] ^ ZOBRIST_CASTLING[0x2] ^ ZOBRIST_CASTLING[0x4];
-    ZOBRIST_CASTLING[0x1 | 0x2 | 0x8] = ZOBRIST_CASTLING[0x1] ^ ZOBRIST_CASTLING[0x2] ^ ZOBRIST_CASTLING[0x8];
-    ZOBRIST_CASTLING[0x1 | 0x4 | 0x8] = ZOBRIST_CASTLING[0x1] ^ ZOBRIST_CASTLING[0x4] ^ ZOBRIST_CASTLING[0x8];
-    ZOBRIST_CASTLING[0x2 | 0x4 | 0x8] = ZOBRIST_CASTLING[0x2] ^ ZOBRIST_CASTLING[0x4] ^ ZOBRIST_CASTLING[0x8];
-    ZOBRIST_CASTLING[0x1 | 0x2 | 0x4 | 0x8] = ZOBRIST_CASTLING[0x1] ^ ZOBRIST_CASTLING[0x2] ^ ZOBRIST_CASTLING[0x4] ^ ZOBRIST_CASTLING[0x8];
-
-    for (int i = 0; i < 8; i++) {
-        ZOBRIST_ENPASSENT[i] = dist(rng);
-    }
-
-    int count = 0;
-    for (Color side = COLOR_WHITE; side <= COLOR_BLACK; side++) {
-        for (Piece piece = 0; piece < PIECE_TYPES; piece++) {
-            for (Square squareA = 0; squareA < 64; squareA++) {
-                for (Square squareB = squareA + 1; squareB < 64; squareB++) {
-                    // Check if there is a reversible move between squareA and squareB
-                    if (piece != PIECE_PAWN && (attackedSquaresByPiece(piece, squareA, C64(0), side) & (C64(1) << squareB))) {
-                        Move move = createMove(squareA, squareB);
-                        uint64_t hash = ZOBRIST_PIECE_SQUARES[side][piece][squareA] ^ ZOBRIST_PIECE_SQUARES[side][piece][squareB] ^ ZOBRIST_STM_BLACK;
-                        int i = H1(hash);
-                        // Find an empty slot in the cuckoo table for this move/hash combination
-                        while (true) {
-                            std::swap(CUCKOO_HASHES[i], hash);
-                            std::swap(CUCKOO_MOVES[i], move);
-                            if (move == 0)
-                                break;
-                            i = (i == H1(hash)) ? H2(hash) : H1(hash);
-                        }
-                        count++;
-                    }
-                }
-            }
-        }
-    }
-    assert(count == 3668);
-}
