+++ src/tt.h
@@ -22,46 +22,6 @@
 #define TT_LOWERBOUND 2
 #define TT_EXACTBOUND 3
 
-extern uint64_t ZOBRIST_PIECE_SQUARES[2][PIECE_TYPES][64];
-extern uint64_t ZOBRIST_STM_BLACK;
-extern uint64_t ZOBRIST_NO_PAWNS;
-extern uint64_t ZOBRIST_CASTLING[16]; // 2^4
-extern uint64_t ZOBRIST_ENPASSENT[8]; // 8 files
-
-extern uint64_t CUCKOO_HASHES[8192];
-extern Move CUCKOO_MOVES[8192];
-
-extern uint8_t TT_GENERATION_COUNTER;
-
-constexpr int H1(uint64_t h) { return h & 0x1fff; }
-constexpr int H2(uint64_t h) { return (h >> 16) & 0x1fff; }
-
-void initZobrist();
-
-struct TTEntry {
-    uint16_t hash = 0;
-    Move bestMove = MOVE_NONE;
-    uint8_t depth = 0;
-    uint8_t flags = TT_NOBOUND;
-    int16_t eval = EVAL_NONE;
-    int16_t value = EVAL_NONE;
-
-    void update(uint64_t _hash, Move _bestMove, uint8_t _depth, Eval _eval, Eval _value, bool wasPv, int _flags) {
-        // Update bestMove if not MOVE_NONE
-        // Or even clear move for a different position
-        if (_bestMove != MOVE_NONE || (uint16_t)_hash != hash)
-            bestMove = _bestMove;
-
-        if (_flags == TT_EXACTBOUND || (uint16_t)_hash != hash || _depth + 2 * wasPv + 4 > depth) {
-            hash = (uint16_t)_hash;
-            depth = _depth;
-            value = _value;
-            eval = _eval;
-            flags = (uint8_t)(_flags + (wasPv << 2)) | TT_GENERATION_COUNTER;
-        }
-    }
-};
-
 inline void* alignedAlloc(size_t alignment, size_t requiredBytes) {
     void* ptr;
 #if defined(__MINGW32__)
@@ -86,6 +46,19 @@ inline void* alignedAlloc(size_t alignment, size_t requiredBytes) {
 #define GENERATION_DELTA (1 << GENERATION_PADDING)
 #define GENERATION_CYCLE 255 + GENERATION_DELTA
 #define GENERATION_MASK (0xFF << GENERATION_PADDING) & 0xFF
+
+extern uint8_t TT_GENERATION_COUNTER;
+
+struct TTEntry {
+    uint16_t hash = 0;
+    Move bestMove = MOVE_NONE;
+    uint8_t depth = NO_DEPTH;
+    uint8_t flags = TT_NOBOUND;
+    int16_t eval = EVAL_NONE;
+    int16_t value = EVAL_NONE;
+
+    void update(uint64_t _hash, Move _bestMove, uint8_t _depth, Eval _eval, Eval _value, bool wasPv, int _flags);
+};
 
 struct TTCluster {
     TTEntry entries[CLUSTER_SIZE];
@@ -133,32 +106,7 @@ public:
         __builtin_prefetch(&table[index(hash)]);
     }
 
-    TTEntry* probe(uint64_t hash, bool* found) {
-        TTCluster* cluster = &table[index(hash)];
-        uint64_t hash16 = (uint16_t)hash;
-
-        TTEntry* replace = &cluster->entries[0];
-
-        for (int i = 0; i < CLUSTER_SIZE; i++) {
-            if (cluster->entries[i].hash == hash16 || !cluster->entries[i].depth) {
-                // Refresh generation
-                cluster->entries[i].flags = (uint8_t)(TT_GENERATION_COUNTER | (cluster->entries[i].flags & (GENERATION_DELTA - 1)));
-                *found = cluster->entries[i].hash == hash16;
-                return &cluster->entries[i];
-            }
-
-            if (i > 0) {
-                // Check if this entry would be better suited for replacement than the current replace entry
-                int replaceValue = replace->depth - ((GENERATION_CYCLE + TT_GENERATION_COUNTER - replace->flags) & GENERATION_MASK);
-                int entryValue = cluster->entries[i].depth - ((GENERATION_CYCLE + TT_GENERATION_COUNTER - cluster->entries[i].flags) & GENERATION_MASK);
-                if (replaceValue > entryValue)
-                    replace = &cluster->entries[i];
-            }
-        }
-
-        *found = false;
-        return replace;
-    }
+    TTEntry* probe(uint64_t hash, bool* found);
 
     int hashfull() {
         int count = 0;
